make percentage double number

+++ src/main/java/com/cob/salesforce/models/dashboard/GenderContainer.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 @Getter
 @Builder
 public class GenderContainer {
-    int malePercentage;
+    double malePercentage;
     int maleNumber;
-    int femalePercentage;
+    double femalePercentage;
     int femaleNumber;
 }
